requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiohttp==3.10.5
beautifulsoup4==4.12.3
python-dotenv==1.0.1
greenlet==3.0.3